requirements.txt
requests>=2.25,<3
python-dotenv>=0.15,<0.16
influxdb>=5.3.1,<5.4
